requirements.txt
Pillow
pdf2image
docx2txt
tesserocr
xlrd
XlsxWriter
numpy
pandas
sklearn
jieba
textwrap3
pinyin